bot_TG/grok_admin_bot/requirements.txt
python-telegram-bot==20.7
requests==2.31.0
python-dotenv==1.0.0
openai==1.3.0
typing-extensions==4.8.0
